fix charset bug on localhost
Who knows if this change did it or if local vscode settings fixed it?
Not me!
Firefox was cached and didn't show me any of my changes! But it's fixed
  now! Yay UTF-8!
Previously, when locally hosting the page with bundle exec jekyll serve,
  redirects - like to 404 errors - were being encoded with Windows-1252
  charset for some reason, even though the html swore it was UTF-8, and
  the original /404 page was fine.  This made arrows, ellipses, and
  other fancy characters show up wrong.
I think moving the charset declaration to the first line of the <head>
  section is what fixed it.